Create and populate default topics.
Adds a column in the users table for each user to have a default
topic, and adds rake tasks to create default topics for each user
that doesn't have one and to move all conversations into those
topics.

Note that topic:default:populate doesn't respect the existence of
any other topics, so it shouldn't be used except on freshly
wiped databases.  Theoretically it shouldn't be too hard to make
it only affect unfiled conversations, but that's an optimization
for another time.