Use gsub instead of replace.
For fucking serious, Ruby.  In what fucked up universe is it okay
to not only have mutable strings, but make the method for
mutating them *which I never ever want to use* take the name of
one of the most common string functions that I want to use
*all the damn time*, and relegate that to a horrible fucking
abomination of a name like gsub.  I fucking hate everything you
stand for.