Don't parse folder models.
Backbone.Collection#set appears to call both Folder#parse and
Folder#initialize on the JSON it gets from the server, even
when models corresponding to that JSON already exist.  I'm not
clear on the whys and wherefores of this, but it's true.  The
upshot is that you shouldn't duplicate work in parse and
initialize.  Parse should be used for transforming the JSON from
the server into the JSON for the model, and initialize is for
transforming the JSON into a Backbone object.