Populate search vector on action save.
This isn't optimal, really.  It adds some extra database work on
every action update, which isn't great.  However, all the other
options I came up with were worse.  We could do it like the
actionvectors rake task which cuts out the select call, but that's
super ugly code (and it's still an extra update).  We could make
a trigger, but that pushes some important logic into the database
which is 1) unlike anything else in the app and 2) not the
ActiveRecord way.

The ideal thing, I think, would be for ActiveRecord to have a way
to assign to a column that doesn't set the value directly, but
tells it how to compute it from the database.  Then we could
set that pretty straightforwardly on message creation and be
done with it.  I don't think that exists, though.